Add VPN support
For clusters running a VPN component to provide site-to-site
connectivity, the current Lokomotive Network Policies are too
restrictive and packets won't reach worker nodes.

This adds the necessary Calico HostEndpoints so GlobalNetworkPolicies
installed by the VPN component can refer to them and allow traffic to
flow.

In particular this adds a HostEndpoint for the tunl0 interfaces of the
workers and HostEndpoints for the vti1 and vti2 devices of worker nodes
that have the "nodetype=vpn" label (indicating they can run the VPN
client).